+++ native/storybook/storybook.js
@@ -54,6 +54,7 @@ configure(() => {
   require('../packages/react-native-bpk-component-banner-alert/stories');
   require('../packages/react-native-bpk-component-button/stories');
   require('../packages/react-native-bpk-component-card/stories');
+  require('../packages/react-native-bpk-component-horizontal-nav/stories');
   require('../packages/react-native-bpk-component-icon/stories');
   require('../packages/react-native-bpk-component-spinner/stories');
   require('../packages/react-native-bpk-component-text-input/stories');